Daily.tsx refactoring and props typing
