+++ frontend-react/src/App.tsx
@@ -1,7 +1,7 @@
 import "./App.css";
 import { Home } from "./pages/Home";
 import { ReportStreamFooter } from "./components/ReportStreamFooter";
-import { Daily } from "./pages/Daily";
+import Daily from "./pages/Daily/Daily";
 import { HowItWorks } from "./pages/HowItWorks";
 import { Details } from "./pages/Details";
 import { Login } from "./pages/Login";
