+++ frontend-react/src/components/HipaaNotice.tsx
@@ -1,0 +1,11 @@
+
+export default function HipaaNotice() {
+    return (
+        <section className="grid-container usa-section usa-prose font-sans-2xs text-base-darker">
+            <p>
+                Data aggregated with ReportStream may be subject to the Privacy Act of 1974, the Health Insurance Portability and Accountability Act of 1996 (HIPAA), and other laws, and requires special safeguarding.
+            </p>
+        </section>
+
+    );
+}
